added a background gradient backround
messed around with the colours of everything and somehow lost the bar along the top with the pages. Tried to add a submit a fun facts thing and failed, or i added it but i want the submit button to send a popup which im failing to get it to do. sorry that the website is so ugly now